P03, P04 and it's tiers buffs, nerfs and changes (#37)
* P04 T2 and 3, Falkok Marksman rifle.

P04 gets 2 upgrades locked behind BT level 3 and 5 and regular falkoks get a precision-based slow fire rate but high damage weapon.

* Add files via upload

* Add files via upload

* Add files via upload

* Add files via upload

* Railgun P03

The railgun unlocks at BT level 4, supposed to be a "power" weapon. Long cooldown but high damage.

* A few Railgun P04 reverts and a fix.

Railgun gets it's old reload time due to testing on 250 HP players that oneshotted them by a body hit if they didnt have armor, BT level requirements lowered by 1 as i thought that level 4 was the score upgrade, also commented it to the reload time and damage.

* Add files via upload

* New helmets, skins and P03 and P04 fixes.

buffs and nerfs for P03 and P04 (the ripper included, more firerate (new values from P04 to T3; 3.2, 4.4 and 4.8) due to players saying that the higher tier rippers were too slow and the P04 outdamaged them and nerfed P04 LMG ROF by .2 most likely not noticable in-game but i thought it would be needed since the gun is almost as fast as the C01-r, 10 more damage to the P03 to make it stronger.