Task rewards progress faster, public task events now work as intended.
Task rewards are now 2x faster to acquire.
Public task events ( sdWeather.EVENT_CRYSTALS_MATTER for example ) now remain active even if the player disconnects, so they can't reconnect a day later and still have this task while it's limited time and can be contributed by all players.
Hopefully no crashes will occur so I don't have to revert this. This event / task is intended to be limited in time so players can choose to complete it or leave it to fade away rather than leaving and completing it later.