Protect SD-BG drone task event
"We are testing old battle drones to see if they are efficient in complementing Star Defenders on this planet. If you would like to be rewarded, make sure it stays intact." - Mothership

An SD task event which summons an old SD Bulky Gunner drone to see if it's performing well on the planet.
Basically put, players have to protect it for 5 minutes, it has 2000 health.