Excavator event, Excavator quartz
An occasional SD event.

Excavator is an entity which digs below it as long as it's powered.
Occasionally the Mothership will send an excavator to the planet
and SD's will need to start it up. The excavator only has power to run for 2 minutes
but players can power it up with Cube shards and Erthal energy cells so it can last longer (1 minute per shard).
They can also repair it with metal shards should it lose health.
	
The excavator eats crystals, gains matter like green BSU does.
After it's excavation process is done - it spits out an unique crystal which always glows white.
The matter of that crystal depends on how much crystals/matter the excavator ate, though the value itself is limited.