More wandering entities
Also wandering entities now indirectly display which factions/mobs can spawn on map
Currently feature is not finished - missing a few factions and bosses but here's how it works:

If Erhtal humanoids/spiders can spawn - an Erthal drone can roam in the sky.
If Falkok faction can spawn - Falkok drones can roam in the sky.

And so on. Some bosses will be roaming the sky too but it's unfinshed at the moment.