New solar panel tier, mothership containers now spawn as events.
New solar panel tier, unlocked at level 36.
These are twice as wide as previous tier, and they're 25% more efficient than two of the previous tier ( 20x instead of 8x multiplier )

Mothership containers
New task type which takes up to a week to complete
Spawns randomly on map, has about 10.4 million matter to fill up. ( 2560 * 5120 matter )
40000 health or similar, not sure. Can survive one or two council bombs until it gets destroyed.
When deployed, mothership will focus on deploying solar matter distributors on the planet, so players can complete those since the distributors now also send matter to the container on the planet if the container exists.
Council attacks the matter container with their bombs occasionally.

The tasks are divided into two phases:
First phase is to protect it and fill it up with mater all the way. Just doing that rewards players that participated in protecting the objective, since the container has a failsafe to teleport back to the mothership if filled with enough matter.
Second phase it to teleport it back to the mothership via long range teleport. Also grants rewards since the container won't have to use failsafe and spend a good portion of it's matter teleporting back to the Mothership, as the LRTP does that for the container.