Setr Scout drone
New drone type for Setr faction.

It has 250 health. When it spots an enemy, it will attempt to fire a flare over the next few seconds, which if done successfully, will summon backup to aid the scout drone.
If it fires a flare successfully, it can't fire another one for about 40 seconds.